apps/conductor/pyproject.toml
[project]
name = "waygate-conductor"
version = "0.1.0"
description = "Conductor service for WayGate"
readme = "README.md"
authors = [
    { name = "Buck Brady", email = "22723438+voidrot@users.noreply.github.com" },
]
requires-python = ">=3.14"
dependencies = [
    "langgraph>=1.1.6",
    "paho-mqtt>=2.1.0",
    "redis>=7.4.0",
    "rq>=2.7.0",
    "waygate-core",
    "waygate-plugin-local-storage",
]

[dependency-groups]
dev = ["pytest>=9.0.3", "pytest-cov>=7.1.0"]

[tool.uv.sources]
waygate-core                 = { workspace = true }
waygate-plugin-local-storage = { workspace = true }

[project.scripts]
waygate-conductor = "waygate_conductor.server:main"

[build-system]
requires      = ["uv_build>=0.11.6,<0.12.0"]
build-backend = "uv_build"

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = [
    "-ra",
    "--strict-config",
    "--strict-markers",
    "--cov=waygate_api",
    "--cov-report=term-missing",
]
